Separate EventXmlHelper into its own EventXmlHelper.scala.
