+++ jvm/core/src/main/scala/org/scalatest/events/Event.scala
@@ -86,120 +86,6 @@ sealed abstract class Event extends Ordered[Event] with Product with Serializabl
 
   // SKIP-SCALATESTJS,NATIVE-START
   private [scalatest] def toXml: Elem
-
-  private[events] object EventXmlHelper {
-    def stringOption(strOption: Option[String]) = strOption.getOrElse("")
-    def longOption(longOption: Option[Long]) = if (longOption.isDefined) longOption.get.toString else ""
-    def booleanOption(booleanOption: Option[Boolean]) = if (booleanOption.isDefined) booleanOption.get.toString else ""
-    def formatterOption(formatterOption: Option[Formatter]) = {
-      formatterOption match {
-        case Some(formatter) =>
-          formatter match {
-            case MotionToSuppress => 
-              <MotionToSuppress/>
-            case indentedText: IndentedText => 
-              <IndentedText>
-                 <formattedText>{ indentedText.formattedText }</formattedText>
-                 <rawText>{ indentedText.rawText }</rawText>
-                 <indentationLevel>{ indentedText.indentationLevel }</indentationLevel>
-              </IndentedText>
-          }
-        case None => ""
-      }
-    }
-    def locationOption(locationOption: Option[Location]) = {
-      locationOption match {
-        case Some(location) =>
-          location match {
-            case topOfClass: TopOfClass =>
-              <TopOfClass>
-                <className>{ topOfClass.className }</className>
-              </TopOfClass>
-            case topOfMethod: TopOfMethod =>
-              <TopOfMethod>
-                <className>{ topOfMethod.className }</className>
-                <methodId>{ topOfMethod.methodId }</methodId>
-              </TopOfMethod>
-            case lineInFile: LineInFile =>
-              <LineInFile>
-                <lineNumber>{ lineInFile.lineNumber }</lineNumber>
-                <fileName>{ lineInFile.fileName }</fileName>
-                <filePathname>{ filePathnameOption(lineInFile.filePathname) }</filePathname>
-              </LineInFile>
-            case SeeStackDepthException =>
-                <SeeStackDepthException />
-            case _ =>
-              ""
-          }
-        case None => ""
-      }
-    }
-    def filePathnameOption(filePathnameOpt: Option[String]) = {
-      filePathnameOpt match {
-        case Some(filePathname) => filePathname
-        case None => ""
-      }
-    }
-    def getThrowableStackDepth(throwable: Throwable) = {
-      throwable match { 
-        case sde: StackDepthException => sde.failedCodeStackDepth 
-        case _ => -1
-      }
-    }
-    def throwableOption(throwableOption: Option[Throwable]) = {
-      throwableOption match {
-        case Some(throwable) => 
-          <message>{ throwable.getMessage }</message>
-          <depth>{ getThrowableStackDepth(throwable) }</depth>
-          <stackTraces>
-            {
-              val stackTraces = throwable.getStackTrace
-              for (stackTrace <- stackTraces) yield {
-                <stackTrace>
-                  <className>{ stackTrace.getClassName }</className>
-                  <methodName>{ stackTrace.getMethodName }</methodName>
-                  <fileName>{ stackTrace.getFileName }</fileName>
-                  <lineNumber>{ stackTrace.getLineNumber }</lineNumber>
-                  <isNative>{ stackTrace.isNativeMethod }</isNative>
-                  <toString>{ stackTrace.toString }</toString>
-                </stackTrace>
-              }
-              /*val stringWriter = new StringWriter()
-              val writer = new PrintWriter(new BufferedWriter(stringWriter))
-              throwable.printStackTrace(writer)
-              writer.flush()
-              stringWriter.toString*/
-            }
-          </stackTraces>
-        case None => ""
-      }
-    }
-    def summaryOption(summaryOption: Option[Summary]) = {
-      summaryOption match {
-        case Some(summary) =>
-          <testsSucceededCount>{ summary.testsSucceededCount }</testsSucceededCount>
-          <testsFailedCount>{ summary.testsFailedCount }</testsFailedCount>
-          <testsIgnoredCount>{ summary.testsIgnoredCount }</testsIgnoredCount>
-          <testsPendingCount>{ summary.testsPendingCount }</testsPendingCount>
-          <testsCanceledCount>{ summary.testsCanceledCount }</testsCanceledCount>
-          <suitesCompletedCount>{ summary.suitesCompletedCount }</suitesCompletedCount>
-          <suitesAbortedCount>{ summary.suitesAbortedCount }</suitesAbortedCount>
-          <scopesPendingCount>{ summary.scopesPendingCount }</scopesPendingCount>
-        case None => ""
-      }
-    }
-    def nameInfoOption(nameInfoOption: Option[NameInfo]) = {
-      nameInfoOption match {
-        case Some(nameInfo) => 
-          <suiteName>{ nameInfo.suiteName }</suiteName>
-          <suiteId>{ nameInfo.suiteId }</suiteId>
-          <suiteClassName>{ stringOption(nameInfo.suiteClassName) }</suiteClassName>
-          <testName>{ stringOption(nameInfo.testName) }</testName>
-        case None => 
-          ""
-      }
-    }
-  }
   // SKIP-SCALATESTJS,NATIVE-END
 
   private[events] object EventJsonHelper {
